requirements_ci.txt
flake8>=3.5.0
coverage>=4.5.1
Sphinx==1.7.1
twine==1.10.0
python-gitlab==1.8.0

pytest
pytest-coverage
pytest-runner
pytest-env

pylint

-r requirements.txt
